requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
httpx==0.24.1
beautifulsoup4==4.12.2
selenium==4.12.0
webdriver-manager==4.0.0
openai==0.28.0
langchain==0.0.268
requests==2.31.0
lxml==4.9.3
python-multipart==0.0.6
pytest==7.4.2
pytest-asyncio==0.21.1
